+++ patches/net/minecraft/world/level/block/ChorusFlowerBlock.java.patch
@@ -5,7 +5,7 @@
          if (p_220981_.isEmptyBlock(blockpos) && blockpos.getY() < p_220981_.getMaxBuildHeight()) {
              int i = p_220980_.getValue(AGE);
 -            if (i < 5) {
-+            if (i < 5 && net.neoforged.neoforge.common.CommonHooks.onCropsGrowPre(p_220981_, blockpos, p_220980_, true)) {
++            if (i < 5 && net.neoforged.neoforge.common.CommonHooks.canCropGrow(p_220981_, blockpos, p_220980_, true)) {
                  boolean flag = false;
                  boolean flag1 = false;
                  BlockState blockstate = p_220981_.getBlockState(p_220982_.below());
@@ -13,7 +13,7 @@
                  } else {
                      this.placeDeadFlower(p_220981_, p_220982_);
                  }
-+                net.neoforged.neoforge.common.CommonHooks.onCropsGrowPost(p_220981_, p_220982_, p_220980_);
++                net.neoforged.neoforge.common.CommonHooks.fireCropGrowPost(p_220981_, p_220982_, p_220980_);
              }
          }
      }
