+++ patches/net/minecraft/world/level/block/CocoaBlock.java.patch
@@ -8,9 +8,9 @@
 +        if (true) {
              int i = p_221000_.getValue(AGE);
 -            if (i < 2) {
-+            if (i < 2 && net.neoforged.neoforge.common.CommonHooks.onCropsGrowPre(p_221001_, p_221002_, p_221000_, p_221001_.random.nextInt(5) == 0)) {
++            if (i < 2 && net.neoforged.neoforge.common.CommonHooks.canCropGrow(p_221001_, p_221002_, p_221000_, p_221001_.random.nextInt(5) == 0)) {
                  p_221001_.setBlock(p_221002_, p_221000_.setValue(AGE, Integer.valueOf(i + 1)), 2);
-+                net.neoforged.neoforge.common.CommonHooks.onCropsGrowPost(p_221001_, p_221002_, p_221000_);
++                net.neoforged.neoforge.common.CommonHooks.fireCropGrowPost(p_221001_, p_221002_, p_221000_);
              }
          }
      }
