+++ patches/net/minecraft/world/level/block/CropBlock.java.patch
@@ -10,9 +10,9 @@
              if (i < this.getMaxAge()) {
                  float f = getGrowthSpeed(this, p_221051_, p_221052_);
 -                if (p_221053_.nextInt((int)(25.0F / f) + 1) == 0) {
-+                if (net.neoforged.neoforge.common.CommonHooks.onCropsGrowPre(p_221051_, p_221052_, p_221050_, p_221053_.nextInt((int)(25.0F / f) + 1) == 0)) {
++                if (net.neoforged.neoforge.common.CommonHooks.canCropGrow(p_221051_, p_221052_, p_221050_, p_221053_.nextInt((int)(25.0F / f) + 1) == 0)) {
                      p_221051_.setBlock(p_221052_, this.getStateForAge(i + 1), 2);
-+                    net.neoforged.neoforge.common.CommonHooks.onCropsGrowPost(p_221051_, p_221052_, p_221050_);
++                    net.neoforged.neoforge.common.CommonHooks.fireCropGrowPost(p_221051_, p_221052_, p_221050_);
                  }
              }
          }
