+++ patches/net/minecraft/world/level/block/GrowingPlantHeadBlock.java.patch
@@ -5,11 +5,11 @@
      @Override
      protected void randomTick(BlockState p_221350_, ServerLevel p_221351_, BlockPos p_221352_, RandomSource p_221353_) {
 -        if (p_221350_.getValue(AGE) < 25 && p_221353_.nextDouble() < this.growPerTickProbability) {
-+        if (p_221350_.getValue(AGE) < 25 && net.neoforged.neoforge.common.CommonHooks.onCropsGrowPre(p_221351_, p_221352_.relative(this.growthDirection), p_221350_, p_221353_.nextDouble() < this.growPerTickProbability)) {
++        if (p_221350_.getValue(AGE) < 25 && net.neoforged.neoforge.common.CommonHooks.canCropGrow(p_221351_, p_221352_.relative(this.growthDirection), p_221350_, p_221353_.nextDouble() < this.growPerTickProbability)) {
              BlockPos blockpos = p_221352_.relative(this.growthDirection);
              if (this.canGrowInto(p_221351_.getBlockState(blockpos))) {
                  p_221351_.setBlockAndUpdate(blockpos, this.getGrowIntoState(p_221350_, p_221351_.random));
-+                net.neoforged.neoforge.common.CommonHooks.onCropsGrowPost(p_221351_, blockpos, p_221351_.getBlockState(blockpos));
++                net.neoforged.neoforge.common.CommonHooks.fireCropGrowPost(p_221351_, blockpos, p_221351_.getBlockState(blockpos));
              }
          }
      }
