+++ patches/net/minecraft/world/level/block/StemBlock.java.patch
@@ -8,7 +8,7 @@
          if (p_222539_.getRawBrightness(p_222540_, 0) >= 9) {
              float f = CropBlock.getGrowthSpeed(this, p_222539_, p_222540_);
 -            if (p_222541_.nextInt((int)(25.0F / f) + 1) == 0) {
-+            if (net.neoforged.neoforge.common.CommonHooks.onCropsGrowPre(p_222539_, p_222540_, p_222538_, p_222541_.nextInt((int)(25.0F / f) + 1) == 0)) {
++            if (net.neoforged.neoforge.common.CommonHooks.canCropGrow(p_222539_, p_222540_, p_222538_, p_222541_.nextInt((int)(25.0F / f) + 1) == 0)) {
                  int i = p_222538_.getValue(AGE);
                  if (i < 7) {
 -                    p_222538_ = p_222538_.setValue(AGE, Integer.valueOf(i + 1));
@@ -27,7 +27,7 @@
                          }
                      }
                  }
-+                net.neoforged.neoforge.common.CommonHooks.onCropsGrowPost(p_222539_, p_222540_, p_222538_);
++                net.neoforged.neoforge.common.CommonHooks.fireCropGrowPost(p_222539_, p_222540_, p_222538_);
              }
          }
      }
