+++ patches/net/minecraft/world/level/block/SugarCaneBlock.java.patch
@@ -13,10 +13,10 @@
  
              if (i < 3) {
                  int j = p_222548_.getValue(AGE);
-+                if (net.neoforged.neoforge.common.CommonHooks.onCropsGrowPre(p_222549_, p_222550_, p_222548_, true)) {
++                if (net.neoforged.neoforge.common.CommonHooks.canCropGrow(p_222549_, p_222550_, p_222548_, true)) {
                  if (j == 15) {
                      p_222549_.setBlockAndUpdate(p_222550_.above(), this.defaultBlockState());
-+                    net.neoforged.neoforge.common.CommonHooks.onCropsGrowPost(p_222549_, p_222550_.above(), this.defaultBlockState());
++                    net.neoforged.neoforge.common.CommonHooks.fireCropGrowPost(p_222549_, p_222550_.above(), this.defaultBlockState());
                      p_222549_.setBlock(p_222550_, p_222548_.setValue(AGE, Integer.valueOf(0)), 4);
                  } else {
                      p_222549_.setBlock(p_222550_, p_222548_.setValue(AGE, Integer.valueOf(j + 1)), 4);
